Investigations: Move Grafana core packages to pkg/registry/apps
