+++ pkg/registry/apps/investigations/authorizer_test.go
@@ -1,4 +1,4 @@
-package app
+package investigations
 
 import (
 	"context"
